# lib/api-zod/src/generated/types/storeReceipt.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * The store's Merkle head and sequence right after a write, and whether the whole chain still verifies. This is what makes an event a receipt rather than a log line.
 */
export interface StoreReceipt {
  head: string;
  seq: number;
  verified: boolean;
}