# lib/api-zod/src/generated/types/studioEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudioEventKind } from './studioEventKind';
import type { StudioEventPayload } from './studioEventPayload';

export interface StudioEvent {
  id: string;
  workspaceId: string;
  kind: StudioEventKind;
  at: Date;
  payload: StudioEventPayload;
  causedBy: string[];
  /** The store's sequence number when this event was written. */
  seq: number;
}