# lib/api-zod/src/generated/types/studioEventInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudioEventInputPayload } from './studioEventInputPayload';
import type { StudioEventKind } from './studioEventKind';

export interface StudioEventInput {
  /** @minLength 1 */
  workspaceId: string;
  kind: StudioEventKind;
  /** When it happened on the operator's clock. Defaults to now. */
  at?: Date;
  /** Kind-specific detail — the ingest endpoint and path for go_live, the exit reason for stream_ended, the source label for source_added. Never a credential. */
  payload?: StudioEventInputPayload;
  /**
     * Ids of earlier events this one follows from.
     * @maxItems 16
     */
  causedBy?: string[];
}