# lib/api-zod/src/generated/types/studioEventKind.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * What happened on air. A closed list; unknown kinds are rejected.
 */
export type StudioEventKind = typeof StudioEventKind[keyof typeof StudioEventKind];


export const StudioEventKind = {
  go_live: 'go_live',
  stream_ended: 'stream_ended',
  scene_saved: 'scene_saved',
  source_added: 'source_added',
  source_removed: 'source_removed',
  stream_error: 'stream_error',
} as const;