# lib/api-zod/src/generated/types/studioEventReceipt.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoreReceipt } from './storeReceipt';
import type { StudioEvent } from './studioEvent';

export interface StudioEventReceipt {
  event: StudioEvent;
  receipt: StoreReceipt;
}