# lib/api-zod/src/generated/types/studioEventTrace.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudioEvent } from './studioEvent';

export interface StudioEventTrace {
  event: StudioEvent;
  /** Transitive causes, nearest first. */
  causes: StudioEvent[];
  /** Direct effects — events that name this one as a cause. */
  effects: StudioEvent[];
}