# lib/api-zod/src/generated/types/studioSceneDocument.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudioSceneDocumentScene } from './studioSceneDocumentScene';

export interface StudioSceneDocument {
  workspaceId: string;
  scene: StudioSceneDocumentScene;
  /** Increments on every save; AS OF on the store can recover any earlier one. */
  version: number;
  savedAt: Date;
  /** The scene_saved event that recorded this version. */
  lastEventId: string;
}