# lib/api-zod/src/generated/types/studioSceneInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudioSceneInputScene } from './studioSceneInputScene';

export interface StudioSceneInput {
  /** The Studio's Scene object — layers with normalised rects. Stored as-is. */
  scene: StudioSceneInputScene;
}