# lib/api-zod/src/generated/types/studioSceneSaved.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoreReceipt } from './storeReceipt';
import type { StudioEvent } from './studioEvent';
import type { StudioSceneDocument } from './studioSceneDocument';

export interface StudioSceneSaved {
  document: StudioSceneDocument;
  event: StudioEvent;
  receipt: StoreReceipt;
}